Moved HapiHL7FileMatcher and HL7FileStream to rs-e2e project given they are not used outside of this project
