+++ rs-e2e/src/main/java/gov/hhs/cdc/trustedintermediary/rse2e/FileFetcher.java
@@ -1,6 +1,5 @@
 package gov.hhs.cdc.trustedintermediary.rse2e;
 
-import gov.hhs.cdc.trustedintermediary.wrappers.HL7FileStream;
 import java.util.List;
 
 /**
