+++ rs-e2e/src/main/java/gov/hhs/cdc/trustedintermediary/rse2e/HL7FileStream.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.wrappers;
+package gov.hhs.cdc.trustedintermediary.rse2e;
 
 import java.io.InputStream;
