+++ rs-e2e/src/main/java/gov/hhs/cdc/trustedintermediary/rse2e/external/hapi/HapiHL7FileMatcher.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.external.hapi;
+package gov.hhs.cdc.trustedintermediary.rse2e.external.hapi;
 
 import ca.uhn.hl7v2.DefaultHapiContext;
 import ca.uhn.hl7v2.HL7Exception;
@@ -6,7 +6,7 @@ import ca.uhn.hl7v2.HapiContext;
 import ca.uhn.hl7v2.model.Message;
 import ca.uhn.hl7v2.model.v251.segment.MSH;
 import ca.uhn.hl7v2.parser.Parser;
-import gov.hhs.cdc.trustedintermediary.wrappers.HL7FileStream;
+import gov.hhs.cdc.trustedintermediary.rse2e.HL7FileStream;
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
 import java.io.IOException;
 import java.io.InputStream;
